[#361] Refactor, move dropdown menu to different files
